backend/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy>=2.0
alembic
asyncpg
python-jose[cryptography]

passlib[bcrypt]
bcrypt==3.2.2

python-multipart
pydantic[email]
psycopg[asyncio]
numpy
pydantic-settings
python-dotenv
